pyproject.toml
[tool.poetry]
name = "ai-perps-trader"
version = "0.1.0"
description = "Multi-agent AI system for perpetuals trading on Hyperliquid via Phantom"
authors = ["Yaowi Kpoga <yaowi@phantom.app>"]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
anthropic = ">=0.40.0"
langchain-anthropic = ">=0.3.0"
langgraph = ">=0.2.0"
langchain-core = ">=0.3.0"
httpx = ">=0.27.0"
python-dotenv = ">=1.0.0"
rich = ">=13.7.0"
pydantic = ">=2.0.0"

[tool.poetry.dev-dependencies]
pytest = ">=8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
